Hide bottom nav bar when opening FiltersController or ReportController. Only hide bottom nav bar in threads when showing bottom snackbar with new posts.

+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/presenter/ThreadPresenter.kt
@@ -1183,7 +1183,7 @@ class ThreadPresenter @Inject constructor(
         val loadable = Loadable.forThread(
           board.site,
           board,
-          threadLink.threadId.toLong(),
+          threadLink.threadId,
           ""
         )
